Move utils functions to utils.py module
* Move updating recheck_auction job to separate util function
