+++ openprocurement/chronograph/scheduler.py
@@ -1,31 +1,24 @@
 # -*- coding: utf-8 -*-
-import grequests
 import requests
-from copy import deepcopy
 from couchdb.http import ResourceConflict
 from datetime import datetime, timedelta
-from gevent.pool import Pool
 from iso8601 import parse_date
 from json import dumps
 from logging import getLogger
-from openprocurement.chronograph.utils import context_unpack
+from openprocurement.chronograph.utils import (
+    context_unpack, update_next_check_job, push, get_request,
+    calc_auction_end_time, get_calendar, skipped_days, get_streams,
+    get_date, find_free_slot, set_date, randomize, get_now, free_slot
+)
 from openprocurement.chronograph.design import plan_auctions_view
 from openprocurement.chronograph.constants import (
-    CALENDAR_ID,
-    STREAMS_ID,
     WORKING_DAY_START,
-    INSIDER_WORKING_DAY_START,
     WORKING_DAY_END,
     WORKING_DAY_DURATION,
-    ROUNDING,
-    MIN_PAUSE,
-    BIDDER_TIME,
-    SERVICE_TIME,
     SMOOTHING_MIN,
     SMOOTHING_REMIN,
     SMOOTHING_MAX,
     NOT_CLASSIC_AUCTIONS,
-    DEFAULT_STREAMS_DOC
 )
 from os import environ
 from pytz import timezone
@@ -42,105 +35,9 @@ ADAPTER = requests.adapters.HTTPAdapter(pool_connections=3, pool_maxsize=3)
 SESSION = requests.Session()
 SESSION.mount('http://', ADAPTER)
 SESSION.mount('https://', ADAPTER)
-POOL = Pool(1)
 
 BASIC_OPT_FIELDS = ['status', 'next_check']
 PLANNING_OPT_FIELDS = ['status', 'next_check', 'auctionPeriod', 'procurementMethodType', 'lots']
-
-
-def get_now():
-    return TZ.localize(datetime.now())
-
-
-def randomize(dt):
-    return dt + timedelta(seconds=randint(0, 1799))
-
-
-def get_calendar(db, calendar_id=CALENDAR_ID):
-    return db.get(calendar_id, {'_id': calendar_id})
-
-
-def set_holiday(db, day):
-    calendar = get_calendar(db)
-    key = parse_date(day).date().isoformat()
-    calendar[key] = True
-    db.save(calendar)
-
-
-def delete_holiday(db, day):
-    calendar = get_calendar(db)
-    key = parse_date(day).date().isoformat()
-    if key in calendar:
-        calendar.pop(key)
-        db.save(calendar)
-
-
-def get_streams(db, streams_id=STREAMS_ID, classic_auction=True):
-        streams = db.get(streams_id, deepcopy(DEFAULT_STREAMS_DOC))
-        if classic_auction:
-            return streams.get('streams', DEFAULT_STREAMS_DOC['streams'])
-        else:
-            return streams.get('dutch_streams',
-                               DEFAULT_STREAMS_DOC['dutch_streams'])
-
-
-def set_streams(db, streams=None, dutch_streams=None, streams_id=STREAMS_ID):
-    streams_doc = db.get(streams_id, deepcopy(DEFAULT_STREAMS_DOC))
-    if streams is not None:
-        streams_doc['streams'] = streams
-    if dutch_streams is not None:
-        streams_doc['dutch_streams'] = dutch_streams
-    db.save(streams_doc)
-
-
-def get_date(db, mode, date, classic_auction=True):
-    plan_id = 'plan{}_{}'.format(mode, date.isoformat())
-    plan = db.get(plan_id, {'_id': plan_id})
-    if classic_auction:
-        plan_date_end = plan.get('time', WORKING_DAY_START.isoformat())
-        stream = plan.get('streams', 1)
-    else:
-        plan_date_end = INSIDER_WORKING_DAY_START.isoformat()
-        stream = len(plan.get('dutch_streams', []))
-    plan_date = parse_date(date.isoformat() + 'T' + plan_date_end, None)
-    plan_date = plan_date.astimezone(TZ) if plan_date.tzinfo else TZ.localize(plan_date)
-    return plan_date.time(), stream, plan
-
-
-def set_date(db, plan, end_time, cur_stream, auction_id, start_time,
-             new_slot=True, classic_auction=True):
-    if classic_auction:
-        if new_slot:
-            plan['time'] = end_time.isoformat()
-            plan['streams'] = cur_stream
-        stream_id = 'stream_{}'.format(cur_stream)
-        stream = plan.get(stream_id, {})
-        stream[start_time.isoformat()] = auction_id
-        plan[stream_id] = stream
-    else:
-        dutch_streams = plan.get('dutch_streams', [])
-        dutch_streams.append(auction_id)
-        plan['dutch_streams'] = dutch_streams
-    db.save(plan)
-
-
-def calc_auction_end_time(bids, start):
-    end = start + bids * BIDDER_TIME + SERVICE_TIME + MIN_PAUSE
-    seconds = (end - TZ.localize(datetime.combine(end, WORKING_DAY_START))).seconds
-    roundTo = ROUNDING.seconds
-    rounding = (seconds + roundTo - 1) // roundTo * roundTo
-    return (end + timedelta(0, rounding - seconds, -end.microsecond)).astimezone(TZ)
-
-
-def find_free_slot(plan):
-    streams = plan.get('streams', 0)
-    for cur_stream in range(1, streams + 1):
-        stream_id = 'stream_{}'.format(cur_stream)
-        for slot in plan[stream_id]:
-            if plan[stream_id].get(slot) is None:
-                plan_date = parse_date(plan['_id'].split('_')[1] + 'T' + slot, None)
-                plan_date = plan_date.astimezone(TZ) if plan_date.tzinfo else TZ.localize(plan_date)
-                return plan_date, cur_stream
 
 
 def planning_auction(auction, start, db, quick=False, lot_id=None):
@@ -199,13 +96,6 @@ def planning_auction(auction, start, db, quick=False, lot_id=None):
     return (start, stream, skipped_days)
 
 
-def skipped_days(days):
-    days_str = ''
-    if days:
-        days_str = ' Skipped {} full days.'.format(days)
-    return days_str
-
-
 def check_auction(request, auction, db):
     now = get_now()
     quick = environ.get('SANDBOX_MODE', False) and u'quick' in auction.get('submissionMethodDetails', '')
@@ -254,36 +144,6 @@ def check_auction(request, auction, db):
     return None
 
 
-def get_request(url, auth, headers=None):
-    tx = ty = 1
-    while True:
-        try:
-            request = grequests.get(url, auth=auth, headers=headers, session=SESSION)
-            grequests.send(request, POOL).join()
-            r = request.response
-        except:
-            pass
-        else:
-            break
-        sleep(tx)
-        tx, ty = ty, tx + ty
-    return r
-
-
-def push(url, params):
-    tx = ty = 1
-    while True:
-        try:
-            r = requests.get(url, params=params)
-        except:
-            pass
-        else:
-            if r.status_code == requests.codes.ok:
-                break
-        sleep(tx)
-        tx, ty = ty, tx + ty
-
-
 def resync_auction(request):
     auction_id = request.matchdict['auction_id']
     scheduler = request.registry.scheduler
@@ -295,13 +155,14 @@ def resync_auction(request):
     request_id = request.environ.get('REQUEST_ID', '')
     next_check = None
     next_sync = None
-    r = get_request(url, auth=(api_token, ''), headers={'X-Client-Request-ID': request_id})
+    r = get_request(
+        url, auth=(api_token, ''), session=SESSION, headers={'X-Client-Request-ID': request_id}
+    )
     if r.status_code != requests.codes.ok:
         LOGGER.error("Error {} on getting auction '{}': {}".format(r.status_code, url, r.text),
                      extra=context_unpack(request, {'MESSAGE_ID': 'error_get_auction'}, {'ERROR_STATUS': r.status_code}))
         if r.status_code == requests.codes.not_found:
             return
-        changes = None
         next_sync = get_now() + timedelta(seconds=randint(SMOOTHING_REMIN, SMOOTHING_MAX))
     else:
         json = r.json()
@@ -318,17 +179,9 @@ def resync_auction(request):
                              extra=context_unpack(request, {'MESSAGE_ID': 'error_patch_auction'}, {'ERROR_STATUS': r.status_code}))
                 next_sync = get_now() + timedelta(seconds=randint(SMOOTHING_REMIN, SMOOTHING_MAX))
             elif r.json():
-                if r.json()['data'].get('next_check'):
-                    next_check = parse_date(r.json()['data']['next_check'], TZ).astimezone(TZ)
+                next_check = r.json()['data'].get('next_check')
     if next_check:
-        check_args = dict(timezone=TZ, id="recheck_{}".format(auction_id),
-                          name="Recheck {}".format(auction_id),
-                          misfire_grace_time=60 * 60, replace_existing=True,
-                          args=[recheck_url, None])
-        if next_check < get_now():
-            scheduler.add_job(push, 'date', run_date=get_now()+timedelta(seconds=randint(SMOOTHING_MIN, SMOOTHING_MAX)), **check_args)
-        else:
-            scheduler.add_job(push, 'date', run_date=next_check+timedelta(seconds=randint(SMOOTHING_MIN, SMOOTHING_MAX)), **check_args)
+        update_next_check_job(next_check, scheduler, auction_id, get_now(), recheck_url)
     if next_sync:
         scheduler.add_job(push, 'date', run_date=next_sync+timedelta(seconds=randint(SMOOTHING_MIN, SMOOTHING_MAX)), timezone=TZ,
                           id=auction_id, name="Resync {}".format(auction_id),
@@ -354,48 +207,11 @@ def recheck_auction(request):
                      extra=context_unpack(request, {'MESSAGE_ID': 'error_check_auction'}, {'ERROR_STATUS': r.status_code}))
         if r.status_code not in [requests.codes.forbidden, requests.codes.not_found]:
             next_check = get_now() + timedelta(minutes=1)
-    elif r.json() and r.json()['data'].get('next_check'):
-        next_check = parse_date(r.json()['data']['next_check'], TZ).astimezone(TZ)
+    elif r.json():
+        next_check = r.json()['data'].get('next_check')
     if next_check:
-        check_args = dict(timezone=TZ, id="recheck_{}".format(auction_id),
-                          name="Recheck {}".format(auction_id),
-                          misfire_grace_time=60 * 60, replace_existing=True,
-                          args=[recheck_url, None])
-        if next_check < get_now():
-            scheduler.add_job(push, 'date', run_date=get_now()+timedelta(seconds=randint(SMOOTHING_MIN, SMOOTHING_MAX)), **check_args)
-        else:
-            scheduler.add_job(push, 'date', run_date=next_check+timedelta(seconds=randint(SMOOTHING_MIN, SMOOTHING_MAX)), **check_args)
+        next_check = update_next_check_job(next_check, scheduler, auction_id, get_now(), recheck_url)
     return next_check and next_check.isoformat()
-
-
-def free_slot(db, plan_id, plan_time, auction_id, classic_auction=True):
-    slot = plan_time.time().isoformat()
-    done = False
-    while not done:
-        try:
-            plan = db.get(plan_id)
-            if classic_auction:
-                streams = plan['streams']
-                for cur_stream in range(1, streams + 1):
-                    stream_id = 'stream_{}'.format(cur_stream)
-                    if plan[stream_id].get(slot) == auction_id:
-                        plan[stream_id][slot] = None
-            else:
-                slots = plan.get('dutch_streams', [])
-                pops = []
-                for i in xrange(0, len(slots)):
-                    if slots[i] == auction_id:
-                        pops.append(i)
-                pops.sort(reverse=True)
-                for p in pops:
-                    slots.pop(p)
-                plan['dutch_streams'] = slots
-            db.save(plan)
-            done = True
-        except ResourceConflict:
-            done = False
-        except:
-            done = True
 
 
 def check_inner_auction(db, auction):
@@ -433,16 +249,8 @@ def process_listing(auctions, scheduler, callback_url, db, check=True, planning=
         tid = auction['id']
         next_check = auction.get('next_check')
         if next_check:
-            check_args = dict(timezone=TZ, id="recheck_{}".format(tid),
-                              name="Recheck {}".format(tid),
-                              misfire_grace_time=60 * 60, replace_existing=True,
-                              args=[callback_url + 'recheck/' + tid, None])
-            next_check = parse_date(next_check, TZ).astimezone(TZ)
-            recheck_job = scheduler.get_job("recheck_{}".format(tid))
-            if next_check < run_date:
-                scheduler.add_job(push, 'date', run_date=run_date+timedelta(seconds=randint(SMOOTHING_MIN, SMOOTHING_MAX)), **check_args)
-            elif not recheck_job or recheck_job.next_run_time != next_check:
-                scheduler.add_job(push, 'date', run_date=next_check+timedelta(seconds=randint(SMOOTHING_MIN, SMOOTHING_MAX)), **check_args)
+            recheck_url = ''.join([callback_url, 'recheck/', tid])
+            update_next_check_job(next_check, scheduler, tid, run_date, recheck_url, True)
         if planning:
             if any([
                 'shouldStartAfter' in i.get('auctionPeriod', {}) and i['auctionPeriod']['shouldStartAfter'] > i['auctionPeriod'].get('startDate')
@@ -461,10 +269,7 @@ def process_listing(auctions, scheduler, callback_url, db, check=True, planning=
 
 def resync_auctions(request):
     next_url = request.params.get('url', '')
-
-    opt_fields = ",".join(BASIC_OPT_FIELDS)
-    if request.registry.planning:
-        opt_fields = ",".join(PLANNING_OPT_FIELDS)
+    opt_fields = ",".join(BASIC_OPT_FIELDS) if not request.registry.planning else ",".join(PLANNING_OPT_FIELDS)
     if not next_url or urlencode({"opt_fields": opt_fields}) not in next_url:
         query = urlencode(OrderedDict(mode='_all_', feed='changes', descending=1, opt_fields=opt_fields))
         next_url = '{}?{}'.format(request.registry.full_url, query)
@@ -474,7 +279,7 @@ def resync_auctions(request):
     request_id = request.environ.get('REQUEST_ID', '')
     while True:
         try:
-            r = get_request(next_url, auth=(api_token, ''), headers={'X-Client-Request-ID': request_id})
+            r = get_request(next_url, auth=(api_token, ''), session=SESSION, headers={'X-Client-Request-ID': request_id})
             if r.status_code == requests.codes.not_found:
                 next_url = ''
                 break
@@ -508,9 +313,7 @@ def resync_auctions(request):
 
 def resync_auctions_back(request):
     next_url = request.params.get('url', '')
-    opt_fields = ",".join(BASIC_OPT_FIELDS)
-    if request.registry.planning:
-        opt_fields = ",".join(PLANNING_OPT_FIELDS)
+    opt_fields = ",".join(BASIC_OPT_FIELDS) if not request.registry.planning else ",".join(PLANNING_OPT_FIELDS)
     if not next_url:
         query = urlencode(OrderedDict(mode='_all_', feed='changes', descending=1, opt_fields=opt_fields))
         next_url = '{}?{}'.format(request.registry.full_url, query)
@@ -521,7 +324,7 @@ def resync_auctions_back(request):
     LOGGER.info("Resync back started", extra=context_unpack(request, {'MESSAGE_ID': 'resync_back_started'}))
     while True:
         try:
-            r = get_request(next_url, auth=(api_token, ''), headers={'X-Client-Request-ID': request_id})
+            r = get_request(next_url, auth=(api_token, ''), session=SESSION, headers={'X-Client-Request-ID': request_id})
             if r.status_code == requests.codes.not_found:
                 next_url = ''
                 break
